Moved types declarations ./dist/@types -> ./src/@types
