+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "JASMINE_CONFIG_PATH=./spec/support/jasmine.json jasmine-ts",
     "build": "npm run build:clean && tsc -p ./tsbuild.json",
-    "build:clean": " rm -rf ./dist/abstracts && rm -rf ./dist/constants && rm -rf ./dist/contracts && rm -rf ./dist/deploy && rm -rf ./dist/helpers && rm -rf ./dist/interfaces && rm -rf ./dist/types",
+    "build:clean": " rm -rf ./dist",
     "build:dry": "tsc -noEmit -p ./tsbuild.json",
     "test-init": "jasmine-ts init"
   },
