+++ spec/helpers/registry.ts
@@ -1,5 +1,5 @@
 import Web3 from 'web3'
-import { Nos } from '../../src/types'
+import { Nos } from '../../src/@types'
 import { increaseTime } from '../../src/helpers'
 import Registry from '../../src/contracts/registry'
 import { ParameterDefaults } from '../../src/constants'
