+++ src/@types/generic.ts
@@ -1,2 +1,3 @@
 export type Nos = number | string
 export type Nosob = number | string | boolean
+
