+++ src/@types/truffle-hdwallet-provider/index.d.ts
