+++ src/contracts/parameterizer.ts
@@ -2,7 +2,7 @@ import Web3 from 'web3'
 import { TransactionReceipt } from 'web3/types.d'
 import { ParameterDefaults, GAS, GAS_PRICE, Errors } from '../constants'
 import Deployable from '../abstracts/deployable'
-import { Nos } from '../types'
+import { Nos } from '../@types'
 import parameterizerJson from '../../computable/build/contracts/Parameterizer.json'
 // TODO PR web3 to export these properly
 import {
