+++ src/interfaces/parameterizer.ts
@@ -2,7 +2,7 @@
  * Define and export interfaces which are integral to the parameterizer contract
  */
 
-import { Nos } from '../types'
+import { Nos } from '../@types'
 
 /**
  * Shape of the parameter object that should be passed to the p11r class during a deploy.
